Add watchdir feature
change settings.cfg.tpl to settings.cfg

+++ Premiumizer.py
@@ -3,6 +3,9 @@
 import os, sys, json
 import time
 import logging
+import win32file
+import win32event
+import win32con
 
 sys.path.insert(1, os.path.abspath(os.path.join(os.path.dirname(__file__), 'lib')))
 
@@ -57,6 +60,12 @@ if prem_config.getboolean('downloads', 'download_enabled'):
     if not os.path.exists(download_path):
         logger.info('Creating Download Path at %s', download_path)
         os.makedirs(download_path)
+        
+if prem_config.getboolean('upload', 'watchdir_enabled'):
+    upload_path = prem_config.get('upload', 'watchdir_location')
+    if not os.path.exists(upload_path):
+        logger.info('Creating Upload Path at %s', upload_path)
+        os.makedirs(upload_path)
 
 logger.info('Initializing Flask')
 app = Flask(__name__)
@@ -84,6 +93,30 @@ class User(UserMixin):
         self.id = userid
         self.password = password
 
+def watchdir():
+    path_to_watch = prem_config.get('upload', 'watchdir_location')
+    change_handle = win32file.FindFirstChangeNotification (
+        path_to_watch,
+        0,
+        win32con.FILE_NOTIFY_CHANGE_FILE_NAME
+    )
+    try:
+        old_path_contents = dict ([(f, None) for f in os.listdir (path_to_watch)])
+        while 1:
+            result = win32event.WaitForSingleObject (change_handle, 500)
+            if result == win32con.WAIT_OBJECT_0:
+                new_path_contents = dict ([(f, None) for f in os.listdir (path_to_watch)])
+                added = [f for f in new_path_contents if not f in old_path_contents]
+                deleted = [f for f in old_path_contents if not f in new_path_contents]
+                if added:
+                    logger.info('Uploading torrent to the cloud: %s', ", ".join (added))
+                    filepath = upload_path + "/" + ''.join(added)
+                    upload_torrent(filepath)
+                    os.remove(filepath)
+                    logger.debug('Deleting torrent from the watchdir: %s', ", ".join (added))
+                old_path_contents = new_path_contents
+                win32file.FindNextChangeNotification (change_handle)
+    finally: win32file.FindCloseChangeNotification (change_handle)
 
 def toUnicode(original, *args):
     try:
@@ -283,7 +316,8 @@ def upload_torrent(filename):
     if response_content['status'] == "success":
         torrents = response_content['torrents']
         parse_tasks(torrents)
-        os.remove(filename)
+        if not prem_config.getboolean('upload', 'watchdir_enabled'):
+            os.remove(filename)
         return True
     else:
         return False
@@ -342,6 +376,10 @@ def settings():
             prem_config.set('downloads', 'download_enabled', 1)
         else:
             prem_config.set('downloads', 'download_enabled', 0)
+        if request.form.get('watchdir_enabled'):
+            prem_config.set('upload', 'watchdir_enabled', 1)
+        else:
+            prem_config.set('upload', 'watchdir_enabled', 0)
         if request.form.get('nzbtomedia_enabled'):
             prem_config.set('nzbtomedia', 'nzbtomedia_enabled', 1)
         else:
@@ -353,6 +391,7 @@ def settings():
         prem_config.set('premiumize', 'pin',  request.form.get('pin'))
         prem_config.set('downloads', 'download_categories',  request.form.get('download_categories'))
         prem_config.set('downloads', 'download_location',  request.form.get('download_location'))
+        prem_config.set('upload', 'watchdir_location',  request.form.get('watchdir_location'))
         prem_config.set('nzbtomedia', 'nzbtomedia_location',  request.form.get('nzbtomedia_location'))
         with open('settings.cfg', 'w') as configfile:    # save
             prem_config.write(configfile)
@@ -448,6 +487,12 @@ def load_tasks():
         task.callback = socketio.emit
         tasks.append(task)
 
+
+# start the watchdir if enabled
+if prem_config.getboolean('upload', 'watchdir_enabled'):
+    t = Thread(target=watchdir)
+    t.daemon = True
+    t.start()
 # start the server with the 'run()' method
 if __name__ == '__main__':
     load_tasks()
@@ -456,4 +501,3 @@ if __name__ == '__main__':
     scheduler.scheduler.add_job(update, 'interval', id='update', seconds=prem_config.getint('global', 'active_interval'), max_instances=1)
     scheduler.start()
     socketio.run(app, port=prem_config.getint('global', 'server_port'))
-
